improved example09

+++ examples/localasset.jl
@@ -2,7 +2,7 @@ function helloWorld2()
   println("HelloWorld1 from localasset")
 end
 
-function squareDiv(a::Float, b::Float)
+function squareDiv(a::Float64, b::Float64)
   println("squareDiv from localasset")
   return (a*a)/(b*b)
 end
